requirements.txt
# Core dependencies for image generation via replicate
replicate

# Image processing
Pillow>=10.0.0

# AWS S3 integration
boto3>=1.34.0

requests

# Python standard library (no installation needed, but listed for reference)
# - os
# - sys
# - logging
# - pathlib
# - typing
# - datetime
# - argparse
# - io